feat: refactor save button in EditProfileForm
