+++ web-components/src/components/buttons/SaveButton.tsx
@@ -1,0 +1,62 @@
+import { makeStyles } from 'tss-react/mui';
+import ContainedButton from 'web-components/src/components/buttons/ContainedButton';
+import { SaveIcon } from 'web-components/src/components/icons/SaveIcon';
+import { Theme } from 'web-components/src/theme/muiTheme';
+
+interface SaveButtonProps {
+  buttonText: string;
+  submitCount: number;
+  isSubmitting: boolean;
+  isValid: boolean;
+  saveDisabled?: boolean;
+}
+
+const useStyles = makeStyles()((theme: Theme) => ({
+  btn: {
+    display: 'flex',
+    minWidth: 0,
+    justifyContent: 'center',
+    [theme.breakpoints.up('sm')]: {
+      fontSize: theme.typography.pxToRem(21),
+      padding: theme.spacing(2, 4),
+      height: theme.typography.pxToRem(60),
+    },
+    [theme.breakpoints.down('sm')]: {
+      fontSize: theme.typography.pxToRem(18),
+      height: theme.typography.pxToRem(50),
+      padding: theme.spacing(2, 3.5),
+    },
+  },
+  saveIcon: {
+    marginRight: theme.spacing(1),
+    height: theme.typography.pxToRem(15),
+    [theme.breakpoints.down('sm')]: {
+      marginBottom: theme.spacing(0.25),
+    },
+  },
+}));
+
+export function SaveButton({
+  buttonText,
+  submitCount,
+  isSubmitting,
+  isValid,
+  saveDisabled,
+}: SaveButtonProps) {
+  const { classes: styles } = useStyles();
+  return (
+    <ContainedButton
+      type="submit"
+      className={styles.btn}
+      disabled={
+        (submitCount > 0 && !isValid) ||
+        isSubmitting ||
+        saveDisabled ||
+        !isValid
+      }
+    >
+      <SaveIcon className={styles.saveIcon} />
+      {buttonText}
+    </ContainedButton>
+  );
+}
